Remove imported infections from intervention page.
- Also moves parse_imported_infections function to entomology page.
- Resolves #28.
